Bug 2029748 - Widen CERT_FormatName length accumulator from unsigned to size_t. r=nss-reviewers,rrelyea
Differential Revision: https://phabricator.services.mozilla.com/D297570

+++ lib/certhigh/certhtml.c
@@ -77,7 +77,8 @@ CERT_FormatName(CERTName *name)
     SECItem *country = 0;
     SECItem *dq = 0;
 
-    unsigned len = 0;
+    /* Use size_t so summing up to 47 unsigned-int field lengths cannot wrap. */
+    size_t len = 0;
     int tag;
     int i;
     int ou_count = 0;
